add nav in component for display data from different point
